Fix schema

+++ main/schemas/user_location.py
@@ -8,9 +8,9 @@ class UserLocationSchema(Schema):
     location_id = ms_fields.Str()
     date_time = ms_fields.DateTime(default=datetime.datetime.now())
     splitted = ms_fields.Bool()
-    accuracy = ms_fields.Int()
-    location_type = ms_fields.Int()
-    longitude = ms_fields.Int()
+    accuracy = ms_fields.Float()
+    location_type = ms_fields.Float()
+    longitude = ms_fields.Float()
     speed = ms_fields.Int()
     arrival = ms_fields.Bool()
